Place code to make Elastic Search request behind an interface so it can be substituted for other implementations.
