+++ Source/ElasticLINQ.Test/ElasticQueryProviderTests.cs
@@ -19,7 +19,7 @@ namespace ElasticLinq.Test
         private static readonly IRetryPolicy retryPolicy = NullRetryPolicy.Instance;
 
         private static readonly ElasticQueryProvider sharedProvider = new ElasticQueryProvider(connection, mapping, log,
-            retryPolicy, "prefix");
+			retryPolicy, "prefix");
 
         private static readonly Expression validExpression =
             Expression.Constant(new ElasticQuery<Sample>(sharedProvider));
@@ -30,7 +30,7 @@ namespace ElasticLinq.Test
         [Fact]
         public void CreateQueryTReturnsElasticQueryTWithProviderSet()
         {
-            var provider = new ElasticQueryProvider(connection, mapping, log, retryPolicy, "prefix");
+			var provider = new ElasticQueryProvider(connection, mapping, log, retryPolicy, "prefix");
 
             var query = provider.CreateQuery<Sample>(validExpression);
 
@@ -41,7 +41,7 @@ namespace ElasticLinq.Test
         [Fact]
         public void CreateQueryReturnsElasticQueryWithProviderSet()
         {
-            var provider = new ElasticQueryProvider(connection, mapping, log, retryPolicy, "prefix");
+			var provider = new ElasticQueryProvider(connection, mapping, log, retryPolicy, "prefix");
 
             var query = provider.CreateQuery(validExpression);
 
@@ -52,7 +52,7 @@ namespace ElasticLinq.Test
         [Fact]
         public void CreateQueryThrowsArgumentOutOfRangeIfExpressionTypeNotAssignableFromIQueryable()
         {
-            var provider = new ElasticQueryProvider(connection, mapping, log, retryPolicy, "prefix");
+			var provider = new ElasticQueryProvider(connection, mapping, log, retryPolicy, "prefix");
 
             Assert.Throws<ArgumentOutOfRangeException>(
                 () => provider.CreateQuery<Sample>(Expression.Constant(new Sample())));
