+++ Source/ElasticLINQ.Test/ElasticQueryTests.cs
@@ -13,7 +13,12 @@ namespace ElasticLinq.Test
     public class ElasticQueryTests
     {
         private static readonly ElasticConnection connection = new ElasticConnection(new Uri("http://localhost"));
-        private static readonly ElasticQueryProvider provider = new ElasticQueryProvider(connection, new TrivialElasticMapping(), NullLog.Instance, NullRetryPolicy.Instance, "prefix");
+		private static readonly ElasticQueryProvider provider = new ElasticQueryProvider(
+			connection,
+			new TrivialElasticMapping(),
+			NullLog.Instance,
+			NullRetryPolicy.Instance,
+			"prefix");
         private static readonly Expression validConstantExpression = Expression.Constant(new ElasticQuery<Sample>(provider));
 
         private class Sample
