+++ Source/ElasticLINQ.Test/Request/Visitors/ElasticQueryTranslation/ElasticQueryTranslationTestsBase.cs
@@ -8,6 +8,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
+using ElasticLinq.Request;
+using NSubstitute;
 
 namespace ElasticLinq.Test.Request.Visitors.ElasticQueryTranslation
 {
