+++ Source/ElasticLINQ/ElasticContext.cs
@@ -5,6 +5,7 @@ using ElasticLinq.Mapping;
 using ElasticLinq.Retry;
 using ElasticLinq.Utility;
 using System.Linq;
+using ElasticLinq.Request;
 
 namespace ElasticLinq
 {
@@ -13,14 +14,14 @@ namespace ElasticLinq
     /// </summary>
     public class ElasticContext : IElasticContext
     {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="ElasticContext"/> class.
-        /// </summary>
-        /// <param name="connection">The information on how to connect to the Elasticsearch server.</param>
-        /// <param name="mapping">The object that helps map queries (optional, defaults to <see cref="TrivialElasticMapping"/>).</param>
-        /// <param name="log">The object which logs information (optional, defaults to <see cref="NullLog"/>).</param>
-        /// <param name="retryPolicy">The object which controls retry policy for the search (optional, defaults to <see cref="RetryPolicy"/>).</param>
-        public ElasticContext(ElasticConnection connection, IElasticMapping mapping = null, ILog log = null, IRetryPolicy retryPolicy = null)
+	    /// <summary>
+	    /// Initializes a new instance of the <see cref="ElasticContext"/> class.
+	    /// </summary>
+	    /// <param name="connection">The information on how to connect to the Elasticsearch server.</param>
+	    /// <param name="mapping">The object that helps map queries (optional, defaults to <see cref="TrivialElasticMapping"/>).</param>
+	    /// <param name="log">The object which logs information (optional, defaults to <see cref="NullLog"/>).</param>
+	    /// <param name="retryPolicy">The object which controls retry policy for the search (optional, defaults to <see cref="RetryPolicy"/>).</param>
+		public ElasticContext(IElasticConnection connection, IElasticMapping mapping = null, ILog log = null, IRetryPolicy retryPolicy = null)
         {
             Argument.EnsureNotNull("connection", connection);
 
@@ -33,7 +34,7 @@ namespace ElasticLinq
         /// <summary>
         /// Specifies the connection to the Elasticsearch server.
         /// </summary>
-        public ElasticConnection Connection { get; private set; }
+		public IElasticConnection Connection { get; private set; }
 
         /// <summary>
         /// The logging mechanism for diagnostic information.
