+++ Source/ElasticLINQ/Request/Facets/IFacet.cs
@@ -4,15 +4,37 @@ using ElasticLinq.Request.Criteria;
 
 namespace ElasticLinq.Request.Facets
 {
-    internal interface IFacet
+    /// <summary>
+	/// Interface that all facets must implement to be part of
+	/// the query facet tree.
+    /// </summary>
+    public interface IFacet
     {
+        /// <summary>
+        /// Name of this facet as specified in the Elasticsearch DSL
+        /// </summary>
         string Name { get; }
+
+        /// <summary>
+		/// Type of this facet as specified in the Elasticsearch DSL
+        /// </summary>
         string Type { get; }
+
+        /// <summary>
+        /// Criteria of this facet
+        /// </summary>
         ICriteria Filter { get; }
     }
 
-    internal interface IOrderableFacet : IFacet
+	/// <summary>
+	/// Interface that all orderable facets must implement to be part of
+	/// the query facet tree.
+	/// </summary>
+	public interface IOrderableFacet : IFacet
     {
+        /// <summary>
+		/// Defines how many top terms should be returned out of the overall terms list
+        /// </summary>
         int? Size { get; }
     }
 }
